+++ Plugin for WP sites/projects/rosenberger/theme/functions.php
@@ -88,7 +88,7 @@ add_filter(
 	function ( $classes ) {
 		if ( is_singular( 'property' ) || is_singular( 'reference' ) || is_post_type_archive( 'reference' )
 				|| is_home() || is_singular( 'post' ) || is_singular( 'region' )
-				|| is_404() || is_page( 'danke' ) ) {
+				|| is_404() || is_page( 'danke' ) || is_page( 'kontakt' ) ) {
 			$classes[] = 'has-light-hero';
 		}
 		return $classes;